+++ src1/apps1/consecutive_name_vol1o0/models/event_volume/ver1o0.py
@@ -12,7 +12,7 @@ class EventVolume(models.Model):
     event_id = models.IntegerField('催事種別Id', blank=True, default=0)
     name = models.CharField('名称', max_length=64, blank=True, null=True)
     name_s = models.CharField('名称_短縮', max_length=16, blank=True, null=True)
-    sort = models.IntegerField('順番', blank=True, default=0)
+    sort = models.FloatField('順番', blank=True, default=0)
 
     def __str__(self):
         """このオブジェクトを文字列にしたとき返るもの"""
